--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "toy-payment-engine"
+name = "toy_payment_engine"
 version = "0.1.0"
 edition = "2018"
 
@@ -11,3 +11,14 @@ debug = true
 [dependencies]
 csv = "1.1"
 serde = { version = "1", features = ["derive"] }
+
+[dev-dependencies]
+criterion = "0.3"
+
+[[bench]]
+name = "benchmark"
+harness = false
+
+[lib]
+name = "lib"
+path = "src/lib.rs"