--- a/Cargo.toml
+++ b/Cargo.toml
@@ -6,7 +6,9 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [profile.release]
-debug = true
+lto = "fat"
+codegen-units = 1
+panic = "abort"
 
 [dependencies]
 csv = "1.1"