--- a/Cargo.toml
+++ b/Cargo.toml
@@ -13,6 +13,7 @@ panic = "abort"
 [dependencies]
 csv = "1.1"
 serde = { version = "1", features = ["derive"] }
+mimalloc = { version = "0.1.17", default-features = false }
 
 [dev-dependencies]
 criterion = "0.3"