--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1,3 +1,6 @@
+#[global_allocator]
+static GLOBAL: mimalloc::MiMalloc = mimalloc::MiMalloc;
+
 use std::io::{self, Write};
 
 pub mod account;