--- a/src/reader.rs
+++ b/src/reader.rs
@@ -1,8 +1,8 @@
 use crate::error::Csv;
 use crate::transaction;
 
-/* 
- * I chose to disgard the CSV entirely if there are elements unparse-able. 
+/*
+ * I chose to disgard the CSV entirely if there are elements unparse-able.
  * One might choose to return a tuple and push back the errors in the console,
  * for further / later manual processing
  * */
@@ -19,7 +19,12 @@ pub fn read_file_from_path(path: &str) -> Result<Vec<transaction::Transaction>, 
             });
 
             if errors.len() > 0 {
-                Err(Csv::ParseError(errors.iter().map(|x| format!("{:?}", x)).collect::<Vec<String>>()))
+                Err(Csv::ParseError(
+                    errors
+                        .iter()
+                        .map(|x| format!("{:?}", x))
+                        .collect::<Vec<String>>(),
+                ))
             } else {
                 Ok(results)
             }