These changes do basic speed improvements by tweaking how it’s being compiled and which allocator is being used

Initial speedup by enabling LTO, change the chunk / crate amount and telling rust not to unwind the stack, but abort on panic;

```bash
main 20                 time:   [226.85 ms 228.21 ms 229.64 ms]
                        change: [-20.141% -19.499% -18.922%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  7 (7.00%) high mild
```

Switching linkers made for another 5%
```bash
main 20                 time:   [220.97 ms 221.91 ms 222.95 ms]
                        change: [-4.9906% -4.3569% -3.7126%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  1 (1.00%) high mild
  12 (12.00%) high severe
```